Edited types.
