+++ components/PlaceCarousel.tsx
@@ -1,8 +1,9 @@
+import { PlaceCardProps } from "@/scripts/types";
 import React, { useRef, useState } from "react";
 import { FlatList, ListRenderItemInfo, View, ViewToken } from "react-native";
 import { cardStyles } from "../assets/styles/card.styles";
 import { placeProfileStyles } from "../assets/styles/place-profile.styles";
-import PlaceCard, { PlaceCardProps } from "./PlaceCard";
+import PlaceCard from "./PlaceCard";
 
 type PlaceCardCarouselProps = {
   places: PlaceCardProps[];
