+++ scripts/config.ts
@@ -1,1 +1,1 @@
-export const BASE_URL: string = "http://192.168.1.102:8080";
+export const BASE_URL: string = "http://192.168.1.106:8080";
